AU10-14 Store post general info in post database draft
Implements post publisher service and database draft for posts. Publisher
service takes a new post from the queue and stores it into a database.
The database uses PostgreSQL as a temporary solution.

For all IDs and sizes changed from uint64 to uint32 as PostgreSQL and
Kafka doesn't have native types uint64 and the size of uint32 is enough for the
nearest future. The external type still is "string".

Fixes spelling errors in names Deal -> Dial.

For database node added new docker image "postdb".